refactor(test): move getChildrenValues test to own file, fix types
- Move job.getChildrenValues() test to flow-children.test.ts (300-line cap)
- Replace `any` with typed assertion for child result values
- Use numeric sort comparator instead of lexicographic
